Move open_archive() to separate module
Also make parameter const and move static functions to anonymous namespace.
