+++ src/meson.build
@@ -13,14 +13,16 @@
 
 main_sources = files('advanced-exif.cc',
 'advanced-exif.h',
+'archives.cc',
+'archives.h',
 'bar.cc',
+'bar.h',
 'bar-comment.cc',
 'bar-comment.h',
 'bar-exif.cc',
 'bar-exif.h',
 'bar-gps.cc',
 'bar-gps.h',
-'bar.h',
 'bar-histogram.cc',
 'bar-histogram.h',
 'bar-keywords.cc',
